backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-phototherapy-backend"
version = "0.1.0"
description = "Backend service and CLI for the VLM phototherapy system."
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.116.0,<0.117.0",
  "google-genai>=1.30.0,<2.0.0",
  "openai>=1.99.0,<2.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "python-multipart>=0.0.20,<0.1.0",
  "sqlalchemy>=2.0.43,<3.0.0",
  "typer>=0.16.0,<0.17.0",
  "uvicorn>=0.35.0,<0.36.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1,<9.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
